+++ src/main/java/io/github/computerdaddyguy/jfiletreeprettyprinter/renderer/EmojiFileFormatter.java
@@ -1,5 +1,6 @@
-package io.github.computerdaddyguy.jfiletreeprettyprinter.renderer.file;
+package io.github.computerdaddyguy.jfiletreeprettyprinter.renderer;
 
+import io.github.computerdaddyguy.jfiletreeprettyprinter.renderer.emoji.EmojiMapping;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.scanner.TreeEntry.DirectoryEntry;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.scanner.TreeEntry.FileEntry;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.scanner.TreeEntry.MaxDepthReachEntry;
@@ -22,7 +23,7 @@ class EmojiFileFormatter implements FileFormatter {
 	}
 
 	private String getFileEmojiPrefix(Path p) {
-		var emoji = emojiMapping.getFileEmoji(p);
+		var emoji = emojiMapping.getPathEmoji(p);
 		return getEmojiPrefix(emoji);
 	}