+++ src/main/java/io/github/computerdaddyguy/jfiletreeprettyprinter/renderer/FileFormatter.java
@@ -1,5 +1,6 @@
-package io.github.computerdaddyguy.jfiletreeprettyprinter.renderer.file;
+package io.github.computerdaddyguy.jfiletreeprettyprinter.renderer;
 
+import io.github.computerdaddyguy.jfiletreeprettyprinter.renderer.emoji.EmojiMapping;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.scanner.TreeEntry.DirectoryEntry;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.scanner.TreeEntry.FileEntry;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.scanner.TreeEntry.MaxDepthReachEntry;
@@ -9,7 +10,7 @@ import java.util.List;
 import org.jspecify.annotations.NullMarked;
 
 @NullMarked
-public interface FileFormatter {
+interface FileFormatter {
 
 	String formatDirectoryBegin(DirectoryEntry dirEntry, List<Path> dirs);
 
@@ -21,10 +22,6 @@ public interface FileFormatter {
 
 	static FileFormatter createDefault() {
 		return new DefaultFileFormatter();
-	}
-
-	static FileFormatter wrapWithEmojis(FileFormatter decorated) {
-		return wrapWithEmojis(decorated, EmojiMapping.createDefault());
 	}
 
 	static FileFormatter wrapWithEmojis(FileFormatter decorated, EmojiMapping emojiMapping) {