+++ src/main/java/io/github/computerdaddyguy/jfiletreeprettyprinter/renderer/LineRenderer.java
@@ -2,7 +2,6 @@ package io.github.computerdaddyguy.jfiletreeprettyprinter.renderer;
 
 import io.github.computerdaddyguy.jfiletreeprettyprinter.renderer.depth.Depth;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.renderer.depth.DepthFormatter;
-import io.github.computerdaddyguy.jfiletreeprettyprinter.renderer.file.FileFormatter;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.scanner.TreeEntry.DirectoryEntry;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.scanner.TreeEntry.FileEntry;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.scanner.TreeEntry.MaxDepthReachEntry;
@@ -31,9 +30,7 @@ interface LineRenderer {
 		var treeFormatter = DepthFormatter.getInstance(options.getTreeFormat());
 
 		var fileFormatter = FileFormatter.createDefault();
-		if (options.areEmojisUsed()) {
-			fileFormatter = FileFormatter.wrapWithEmojis(fileFormatter);
-		}
+		fileFormatter = FileFormatter.wrapWithEmojis(fileFormatter, options.getEmojiMapping());
 
 		return new DefaultLineRenderer(treeFormatter, fileFormatter);
 	}