+++ src/test/java/io/github/computerdaddyguy/jfiletreeprettyprinter/EmojisTest.java
@@ -2,8 +2,11 @@ package io.github.computerdaddyguy.jfiletreeprettyprinter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import io.github.computerdaddyguy.jfiletreeprettyprinter.renderer.emoji.EmojiMapping;
+import io.github.computerdaddyguy.jfiletreeprettyprinter.util.FileStructureCreator;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.util.FileStructures;
 import java.nio.file.Path;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
@@ -12,14 +15,15 @@ class EmojisTest {
 	@TempDir
 	private Path root;
 
-	private FileTreePrettyPrinter printer = FileTreePrettyPrinter.builder()
-		.customizeOptions(
-			options -> options.withEmojis(true)
-		)
-		.build();
-
 	@Test
 	void emptyDir() {
+
+		var printer = FileTreePrettyPrinter.builder()
+			.customizeOptions(
+				options -> options.withDefaultEmojis()
+			)
+			.build();
+
 		var path = FileStructures.emptyDirectory(root);
 		var result = printer.prettyPrint(path);
 		var expected = "📂 targetPath/";
@@ -28,6 +32,12 @@ class EmojisTest {
 
 	@Test
 	void emojis() {
+
+		var printer = FileTreePrettyPrinter.builder()
+			.customizeOptions(
+				options -> options.withDefaultEmojis()
+			)
+			.build();
 
 		var result = printer.prettyPrint("src/example/resources/emojis");
 		var expected = """
@@ -155,6 +165,191 @@ class EmojisTest {
 			   └─ 🗑️ file.tmp""";
 
 		assertThat(result).isEqualTo(expected);
+	}
+
+	@Nested
+	class DirectoryEmojiMapping {
+
+		@Test
+		void dir_name() {
+
+			// @formatter:off
+			var path = FileStructureCreator
+				.forTargetPath(root)
+				.createDirectory("dirA")
+				.createDirectory("dirB")
+				.createDirectory("dirC")
+				.getPath();
+			// @formatter:on
+
+			var mapping = EmojiMapping.builderFromDefault()
+				.setDirectoryNameEmoji("dirA", "⭐") // add emoji 
+				.build();
+
+			var printer = FileTreePrettyPrinter.builder()
+				.customizeOptions(
+					options -> options.withEmojis(mapping)
+				)
+				.build();
+
+			var result = printer.prettyPrint(path);
+
+			var expected = """
+				📂 targetPath/
+				├─ ⭐ dirA/
+				├─ 📂 dirB/
+				└─ 📂 dirC/""";
+
+			assertThat(result).isEqualTo(expected);
+		}
+
+		@Test
+		void dir_match() {
+
+			// @formatter:off
+			var path = FileStructureCreator
+				.forTargetPath(root)
+				.createDirectory("dirA")
+				.createDirectory("dirB")
+				.createDirectory("dirC")
+				.getPath();
+			// @formatter:on
+
+			var mapping = EmojiMapping.builderFromDefault()
+				.addDirectoryEmoji(PathMatchers.hasName("dirA"), "⭐") // add emoji
+				.addDirectoryEmoji(PathMatchers.hasName("dirB"), "😊") // change existing emoji
+				.build();
+
+			var printer = FileTreePrettyPrinter.builder()
+				.customizeOptions(
+					options -> options.withEmojis(mapping)
+				)
+				.build();
+
+			var result = printer.prettyPrint(path);
+
+			var expected = """
+				📂 targetPath/
+				├─ ⭐ dirA/
+				├─ 😊 dirB/
+				└─ 📂 dirC/""";
+
+			assertThat(result).isEqualTo(expected);
+		}
+
+	}
+
+	@Nested
+	class FileEmojiMapping {
+
+		@Test
+		void file_name() {
+
+			// @formatter:off
+			var path = FileStructureCreator
+				.forTargetPath(root)
+				.createFile("aaa") 
+				.createFile("dockerfile") 
+				.createFile("jenkinsfile") 
+				.createFile("license") 
+				.getPath();
+			// @formatter:on
+
+			var mapping = EmojiMapping.builderFromDefault()
+				.setFileNameEmoji("aaa", "⭐") // add emoji 
+				.setFileNameEmoji("dockerfile", "😊") // change existing emoji
+				.build();
+
+			var printer = FileTreePrettyPrinter.builder()
+				.customizeOptions(
+					options -> options.withEmojis(mapping)
+				)
+				.build();
+
+			var result = printer.prettyPrint(path);
+
+			var expected = """
+				📂 targetPath/
+				├─ ⭐ aaa
+				├─ 😊 dockerfile
+				├─ 🤵 jenkinsfile
+				└─ ⚖️ license""";
+
+			assertThat(result).isEqualTo(expected);
+		}
+
+		@Test
+		void file_extension() {
+
+			// @formatter:off
+			var path = FileStructureCreator
+				.forTargetPath(root)
+				.createFile("file.plop") 
+				.createFile("file.avi") 
+				.createFile("file.gif") 
+				.createFile("license") 
+				.getPath();
+			// @formatter:on
+
+			var mapping = EmojiMapping.builderFromDefault()
+				.setFileExtensionEmoji("plop", "⭐") // add emoji
+				.setFileExtensionEmoji("avi", "😊") // change existing emoji
+				.build();
+
+			var printer = FileTreePrettyPrinter.builder()
+				.customizeOptions(
+					options -> options.withEmojis(mapping)
+				)
+				.build();
+
+			var result = printer.prettyPrint(path);
+
+			var expected = """
+				📂 targetPath/
+				├─ 😊 file.avi
+				├─ 🎞️ file.gif
+				├─ ⭐ file.plop
+				└─ ⚖️ license""";
+
+			assertThat(result).isEqualTo(expected);
+		}
+
+		@Test
+		void file_match() {
+
+			// @formatter:off
+			var path = FileStructureCreator
+				.forTargetPath(root)
+				.createFile("file.plop") 
+				.createFile("file.avi") 
+				.createFile("file.gif") 
+				.createFile("license") 
+				.getPath();
+			// @formatter:on
+
+			var mapping = EmojiMapping.builderFromDefault()
+				.addFileEmoji(PathMatchers.hasName("file.plop"), "⭐") // add emoji
+				.addFileEmoji(PathMatchers.hasName("file.avi"), "😊") // change existing emoji
+				.build();
+
+			var printer = FileTreePrettyPrinter.builder()
+				.customizeOptions(
+					options -> options.withEmojis(mapping)
+				)
+				.build();
+
+			var result = printer.prettyPrint(path);
+
+			var expected = """
+				📂 targetPath/
+				├─ 😊 file.avi
+				├─ 🎞️ file.gif
+				├─ ⭐ file.plop
+				└─ ⚖️ license""";
+
+			assertThat(result).isEqualTo(expected);
+		}
+
 	}
 
 }