+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ---
 ## [0.0.6] - Unreleased
 
+### Added
+- Emojis: option to use custom mapping
+
 ### Changed
 - Child limit: do not print skipped children count
 - Max depth: do not print "max depth reached"